refactor(compose): the dedupe adapter takes its own file
attentionseam.go crossed the 500-line cap once this branch's
SettleablePairs landed on top of what main had grown.

The dedupe adapter is the one concept in the seam that asks the store
two different questions about the same page — which pairs are open, and
which of those a merge would accept. Everything else the seam binds is a
single read, so it is the split the file was already suggesting rather
than a cut made to reach a number.
